chức năng tìm kiếm, xóa xử lý

+++ src/main/java/qltv/web/controllers/XuLyController.java
@@ -5,6 +5,8 @@
 package qltv.web.controllers;
 
 import jakarta.validation.Valid;
+
+import java.util.Date;
 import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
@@ -15,11 +17,15 @@ import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.ResponseBody;
 import qltv.web.dto.ThanhVienDTO;
 import qltv.web.dto.XuLyDTO;
+import qltv.web.mappers.ThanhVienMapper;
+import qltv.web.models.ThanhVien;
 import qltv.web.security.SecurityUtil;
 import qltv.web.services.ThanhVienService;
 import qltv.web.services.XuLyService;
+import qltv.web.services.impl.ThanhVienServiceImpl;
 
 @Controller
 public class XuLyController {
@@ -42,14 +48,14 @@ public class XuLyController {
         Long maTV = Long.parseLong(username);
         ThanhVienDTO user = thanhVienService.findMemberById(maTV);
         List<XuLyDTO> xuLys = xuLyService.getAllXuLy();
-        
+
         model.addAttribute("user", user);
         model.addAttribute("xuLys", xuLys);
         return "xu-ly-list";
     }
-    
+
     @GetMapping("/xuly/new")
-    public String CreateXuLy(Model model){
+    public String CreateXuLy(Model model) {
         String username = SecurityUtil.getUserSession();
         if (username == null) { // kiểm tra đăng nhập
             return "redirect:/login";
@@ -57,24 +63,68 @@ public class XuLyController {
         Long maTV = Long.parseLong(username);
         ThanhVienDTO user = thanhVienService.findMemberById(maTV);
         XuLyDTO xuLy = new XuLyDTO();
-        
+        Long nextMaXL = xuLyService.getMaxMaXL() + 1;
+        List<ThanhVienDTO> thanhViens = thanhVienService.getAllThanhVien();
+
         model.addAttribute("user", user);
         model.addAttribute("xuLy", xuLy);
+        model.addAttribute("nextMaXL", nextMaXL);
+        model.addAttribute("thanhViens", thanhViens);
+
         return "xu-ly-create";
     }
-    
+
     @PostMapping("/xuly/new")
-    public String saveXuLy(@Valid @ModelAttribute("xuLy") XuLyDTO xuLy, BindingResult result,
+    public String saveXuLy(@RequestParam("maXuLy") String maXuLy,
+            @RequestParam("maThanhVien") String maThanhVien,
+            @RequestParam("hinhThucXuLy") String hinhThucXuLy,
+            @RequestParam("tienPhat") String tienPhat,
+            @RequestParam("trangThai") String trangThai,
             Model model) {
-        if (result.hasErrors()) {
-            model.addAttribute("xuLy", xuLy);
+        String username = SecurityUtil.getUserSession();
+        if (username == null) {
+            return "redirect:/login";
+        }
+
+        String validate = xuLyService.validateXuLy(maThanhVien, hinhThucXuLy, tienPhat);
+        if (!validate.isEmpty()) {
+            model.addAttribute("error", validate);
             return "xu-ly-create";
         }
+
+        switch (hinhThucXuLy) {
+            case "KhoaThe2Thang":
+                hinhThucXuLy = "Khóa thẻ 2 tháng";
+                break;
+            case "KhoaThe3Thang":
+                hinhThucXuLy = "Khóa thẻ 3 tháng";
+                break;
+            case "KhoaThe1ThangVaBoiThuong":
+                hinhThucXuLy = "Khóa thẻ 1 tháng và bồi thường";
+                break;
+            case "BoiThuong":
+                hinhThucXuLy = "Bồi thường";
+                break;
+            default:
+                hinhThucXuLy = "Khóa thẻ 1 tháng";
+        }
+
+        if (tienPhat.isEmpty()) {
+            tienPhat = "0";
+        }
+
+        XuLyDTO xuLy = new XuLyDTO();
+        xuLy.setMaXL(Integer.parseInt(maXuLy));
+        xuLy.setThanhVien(thanhVienService.findMemberById(Integer.parseInt(maThanhVien)));
+        xuLy.setHinhThucXL(hinhThucXuLy);
+        xuLy.setSoTien(Integer.parseInt(tienPhat));
+        xuLy.setTrangThaiXL(Integer.parseInt(trangThai));
+        xuLy.setNgayXL(new Date());
 
         xuLyService.saveXuLy(xuLy);
         return "redirect:/xuly";
     }
-    
+
     @GetMapping("/xuly/{maXL}/edit")
     public String editXuLy(@PathVariable("maXL") long maXL, Model model) {
         String username = SecurityUtil.getUserSession();
@@ -84,17 +134,17 @@ public class XuLyController {
         Long maTV = Long.parseLong(username);
         ThanhVienDTO user = thanhVienService.findMemberById(maTV);
         XuLyDTO xuLy = xuLyService.findXuLyByMaXL(maXL);
-        
+
         model.addAttribute("user", user);
         model.addAttribute("xuLy", xuLy);
-        
+
         return "xu-ly-edit";
     }
 
     @PostMapping("/xuly/{maXL}/edit")
     public String updateXuLy(@PathVariable("maXL") long maXL,
-                               @Valid @ModelAttribute("xuLy") XuLyDTO xuLy,
-                                BindingResult result, Model model) {
+            @Valid @ModelAttribute("xuLy") XuLyDTO xuLy,
+            BindingResult result, Model model) {
 
         String username = SecurityUtil.getUserSession();
         if (username == null) {
@@ -118,25 +168,23 @@ public class XuLyController {
         if (username == null) {
             return "redirect:/login";
         }
-        
-        xuLyService.deleteXuLy(maXL);
+
+        if (xuLyService.findXuLyByMaXL(maXL) != null) {
+            xuLyService.deleteXuLy(maXL);
+        }
         return "redirect:/xuly";
     }
 
     @GetMapping("/xuly/search")
-    public String searchThanhVien(@RequestParam(value = "query") String query, Model model) {
+    @ResponseBody
+    public List<XuLyDTO> searchThanhVien(@RequestParam(value = "query") String query, Model model) {
         String username = SecurityUtil.getUserSession();
         if (username == null) {
-            return "redirect:/login";
+            return null;
         }
         Long maTV = Long.parseLong(username);
         ThanhVienDTO user = thanhVienService.findMemberById(maTV);
         List<XuLyDTO> xuLys = xuLyService.searchXuLy(query);
-        
-        model.addAttribute("user", user);
-        model.addAttribute("xuLys", xuLys);
-        return "xu-ly-list";
+        return xuLys;
     }
-    
-    
 }
